server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
Pillow>=10.0.0
open3d>=0.18.0
trimesh>=4.0.0
numpy>=1.24.0
